requirements.txt
fastapi==0.103.2
pydantic-settings==2.0.3
uvicorn==0.23.2 # waitress package similar to it 
cassandra-driver==3.28.0
python-dotenv==1.0.0
email-validator==2.0.0.post2
argon2-cffi==23.1.0
pytest==7.4.2
# jinja2 # no need already built in fastapi module
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
urlparse